refactor: Movement 검증 로직 이동
