+++ src/main/java/domain/piece/path/FixedMultiStepPathFinder.java
@@ -1,6 +1,8 @@
 package domain.piece.path;
 
+import domain.position.Direction;
 import domain.position.Position;
+import java.util.ArrayList;
 import java.util.List;
 
 public class FixedMultiStepPathFinder implements PathFinder {
@@ -12,10 +14,38 @@ public class FixedMultiStepPathFinder implements PathFinder {
 
     @Override
     public List<Position> findIntermediatePositions(Position from, Position to) {
-        return movements.stream()
-                .filter(movement -> movement.isValidMove(from, to))
-                .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("해당 좌표로 이동시킬 수 없습니다."))
-                .findIntermediatePositions(from, to);
+        List<Direction> finMovement = movements.stream()
+                .map(Movement::getMovement)
+                .filter(movement -> isValidMove(from, to, movement))
+                .findAny()
+                .orElseThrow(() -> new IllegalArgumentException("해당 좌표로 이동시킬 수 없습니다."));
+        return getPathPositionsFrom(from, finMovement);
+    }
+
+    private boolean isValidMove(Position startPosition, Position endPosition, List<Direction> movement) {
+        Position curPosition = startPosition;
+        for (Direction direction : movement) {
+            if (isInvalidMove(curPosition, direction)) {
+                return false;
+            }
+            curPosition = curPosition.movePosition(direction.getDeltaRow(), direction.getDeltaColumn());
+        }
+        return curPosition.equals(endPosition);
+    }
+
+    private List<Position> getPathPositionsFrom(Position startPosition, List<Direction> movement) {
+        List<Position> pathPositions = new ArrayList<>();
+
+        for (Direction direction : movement) {
+            startPosition = startPosition.movePosition(direction.getDeltaRow(), direction.getDeltaColumn());
+            pathPositions.add(startPosition);
+        }
+
+        pathPositions.removeLast();
+        return pathPositions;
+    }
+
+    private boolean isInvalidMove(Position curPosition, Direction direction) {
+        return !curPosition.canMovePosition(direction.getDeltaRow(), direction.getDeltaColumn());
     }
 }
